[transaction builder generator] share more code with common::get_required_decoding_helper_types
