+++ language/transaction-builder/generator/src/python3.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::common::{mangle_type, type_not_allowed};
+use crate::common;
 use heck::{CamelCase, ShoutySnakeCase};
 use libra_types::transaction::{ArgumentABI, ScriptABI, TypeArgumentABI};
 use move_core_types::language_storage::TypeTag;
@@ -11,7 +11,7 @@ use serde_generate::{
 };
 
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeMap,
     io::{Result, Write},
     path::PathBuf,
 };
@@ -208,7 +208,7 @@ def decode_script(script: Script) -> ScriptCall:
                 self.out,
                 "{}=decode_{}_argument(script.args[{}]),",
                 arg.name(),
-                mangle_type(arg.type_tag()),
+                common::mangle_type(arg.type_tag()),
                 index,
             )?;
         }
@@ -257,21 +257,7 @@ SCRIPT_ENCODER_MAP: typing.Dict[typing.Type[ScriptCall], typing.Callable[[Script
     }
 
     fn output_decoding_helpers(&mut self, abis: &[ScriptABI]) -> Result<()> {
-        let mut required_types = BTreeSet::new();
-        for abi in abis {
-            for arg in abi.args() {
-                let mut type_tag = arg.type_tag();
-                required_types.insert(type_tag);
-                while let TypeTag::Vector(inner_type_tag) = type_tag {
-                    if **inner_type_tag == TypeTag::U8 {
-                        // decode_u8vector does not call decode_u8.
-                        break;
-                    }
-                    type_tag = inner_type_tag;
-                    required_types.insert(type_tag);
-                }
-            }
-        }
+        let required_types = common::get_required_decoding_helper_types(abis);
         for required_type in required_types {
             self.output_decoding_helper(required_type)?;
         }
@@ -292,11 +278,11 @@ SCRIPT_ENCODER_MAP: typing.Dict[typing.Type[ScriptCall], typing.Callable[[Script
                     "Vector",
                     format!(
                         "[decode_{}_argument(x) for x in arg.value]",
-                        mangle_type(inner_type_tag)
+                        common::mangle_type(inner_type_tag)
                     ),
                 ),
             },
-            Struct(_) | Signer => type_not_allowed(type_tag),
+            Struct(_) | Signer => common::type_not_allowed(type_tag),
         };
         writeln!(
             self.out,
@@ -306,7 +292,7 @@ def decode_{}_argument(arg: TransactionArgument) -> {}:
         raise ValueError("Was expecting a {} argument")
     return {}
 "#,
-            mangle_type(type_tag),
+            common::mangle_type(type_tag),
             Self::quote_type(type_tag),
             constructor,
             constructor,
@@ -375,10 +361,10 @@ def decode_{}_argument(arg: TransactionArgument) -> {}:
             Address => "AccountAddress".into(),
             Vector(type_tag) => match type_tag.as_ref() {
                 U8 => "bytes".into(),
-                _ => type_not_allowed(type_tag),
+                _ => common::type_not_allowed(type_tag),
             },
 
-            Struct(_) | Signer => type_not_allowed(type_tag),
+            Struct(_) | Signer => common::type_not_allowed(type_tag),
         }
     }
 
@@ -392,10 +378,10 @@ def decode_{}_argument(arg: TransactionArgument) -> {}:
             Address => format!("TransactionArgument__Address(value={})", name),
             Vector(type_tag) => match type_tag.as_ref() {
                 U8 => format!("TransactionArgument__U8Vector(value={})", name),
-                _ => type_not_allowed(type_tag),
+                _ => common::type_not_allowed(type_tag),
             },
 
-            Struct(_) | Signer => type_not_allowed(type_tag),
+            Struct(_) | Signer => common::type_not_allowed(type_tag),
         }
     }
 }
